[FIX] Correct positions of P3 and P4
At some time prior to the previous commit, the pogo pin footprints P3
and P4 got moved slightly. This was likely done accidentally, e.g., when
changing the grid spacing. As of this commit, the positions of `P3` and
`P4` have been corrected.

N.B., as part of the position correction corrections, the accompanying
traces were disconnected (perhaps there's a way to move the traces
precisely without disconnecting the traces from the pads?).  As part of
this commit, the disconnected traces were manually reconnected (should
be fine, but something might have gotten messed up so it is worth
documenting here).